Add files via upload
When we start up many devices, often an LED blinks to indicate that the devices are booting up or that it has been switched on. And at the middle of the night I just wondering how to code the hardware. 

As part of my exploration into embedded systems, I tried writing a program on my STM 32 Nucleo board to blink an LED once upon boot up. The objective was simple for me to use the STM32 HAL library to configure GPIO and EXTI (External Interrupt) for a simple yet effective demonstration. I initialized the GPIO pin connected to the LED as an output and set up another GPIO pin for external interrupt, configured to trigger on a rising edge. By developing an Interrupt Service Routine (ISR), I ensured the LED toggled and the EXTI interrupt flag was cleared correctly. Additionally, I configured the system clock using CubeMX to maintain precise timing and reliable operation. This project provided hands-on experience in handling external interruptions and knowing about the Interrupt priority register (IPR).